requirements.txt
python-telegram-bot==13.14
python-dotenv==0.19.2
gspread==5.7.1
oauth2client==4.1.3
gspread-formatting==1.0.9
PyDrive2==1.10.0
